refactor(*): switch to github.com/alebabai/go-kafka
