+++ examples/common/consumer/request_response.go
@@ -1,7 +1,7 @@
 package consumer
 
 import (
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
 )
 
 type CreateEventRequest struct {
