+++ examples/common/consumer/storage.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/go-kit/log"
 
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
 )
 
 type StorageService struct {
