+++ examples/common/producer/generator.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-kit/log"
 	"github.com/google/uuid"
 
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
 )
 
 type GeneratorService struct {
