+++ examples/common/producer/service.go
@@ -3,7 +3,7 @@ package producer
 import (
 	"context"
 
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
 )
 
 type Service interface {
