+++ examples/confluent/cmd/consumer/main.go
@@ -8,16 +8,16 @@ import (
 	"os/signal"
 	"syscall"
 
-	kafkatransport "github.com/alebabai/go-kit-kafka/kafka/transport"
+	kafkatransport "github.com/alebabai/go-kit-kafka/v2/transport"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/go-kit/kit/transport"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 
-	"github.com/alebabai/go-kit-kafka/examples/common/consumer"
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/consumer"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
 
-	"github.com/alebabai/go-kit-kafka/examples/confluent/pkg/consumer/kafka/adapter"
+	"github.com/alebabai/go-kit-kafka/v2/examples/confluent/pkg/consumer/kafka/adapter"
 )
 
 func fatal(logger log.Logger, err error) {
