+++ examples/confluent/go.mod
@@ -1,15 +1,22 @@
-module github.com/alebabai/go-kit-kafka/examples/confluent
+module github.com/alebabai/go-kit-kafka/v2/examples/confluent
 
-go 1.16
+go 1.22
 
 require (
-	github.com/alebabai/go-kit-kafka v0.0.0
-	github.com/alebabai/go-kit-kafka/examples/common v0.0.0
+	github.com/alebabai/go-kafka v0.3.0
+	github.com/alebabai/go-kit-kafka/v2 v2.0.0
+	github.com/alebabai/go-kit-kafka/v2/examples/common v0.0.0
 	github.com/confluentinc/confluent-kafka-go v1.8.2
 	github.com/go-kit/kit v0.13.0
 	github.com/go-kit/log v0.2.1
 )
 
-replace github.com/alebabai/go-kit-kafka => ../..
+require (
+	github.com/go-logfmt/logfmt v0.5.1 // indirect
+	github.com/google/uuid v1.2.0 // indirect
+	github.com/gorilla/mux v1.8.0 // indirect
+)
 
-replace github.com/alebabai/go-kit-kafka/examples/common => ../common
+replace github.com/alebabai/go-kit-kafka/v2 => ../..
+
+replace github.com/alebabai/go-kit-kafka/v2/examples/common => ../common
