+++ examples/confluent/pkg/consumer/kafka/adapter/listener.go
@@ -7,7 +7,7 @@ import (
 	"io"
 	"time"
 
-	kitkafka "github.com/alebabai/go-kit-kafka/kafka"
+	kitkafka "github.com/alebabai/go-kafka"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/go-kit/kit/transport"
 )
