+++ examples/confluent/pkg/consumer/kafka/adapter/message.go
@@ -1,7 +1,7 @@
 package adapter
 
 import (
-	kitkafka "github.com/alebabai/go-kit-kafka/kafka"
+	kitkafka "github.com/alebabai/go-kafka"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 )
