+++ examples/sarama/cmd/producer/main.go
@@ -14,10 +14,10 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 
-	"github.com/alebabai/go-kit-kafka/examples/common/domain"
-	"github.com/alebabai/go-kit-kafka/examples/common/producer"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/domain"
+	"github.com/alebabai/go-kit-kafka/v2/examples/common/producer"
 
-	"github.com/alebabai/go-kit-kafka/examples/sarama/pkg/producer/kafka/adapter"
+	"github.com/alebabai/go-kit-kafka/v2/examples/sarama/pkg/producer/kafka/adapter"
 )
 
 func fatal(logger log.Logger, err error) {
