+++ examples/sarama/pkg/consumer/kafka/adapter/consumer_group_handler.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/Shopify/sarama"
-	"github.com/alebabai/go-kit-kafka/kafka"
+	"github.com/alebabai/go-kafka"
 	"github.com/go-kit/kit/transport"
 )
