+++ go.mod
@@ -1,9 +1,17 @@
-module github.com/alebabai/go-kit-kafka
+module github.com/alebabai/go-kit-kafka/v2
 
-go 1.16
+go 1.22
 
 require (
+	github.com/alebabai/go-kafka v0.3.0
 	github.com/go-kit/kit v0.13.0
 	github.com/go-kit/log v0.2.1
 	github.com/opentracing/opentracing-go v1.2.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-logfmt/logfmt v0.5.1 // indirect
+	github.com/stretchr/testify v1.7.2 // indirect
+	google.golang.org/grpc v1.40.0 // indirect
 )
