--- kafka/doc.go
@@ -1,2 +1,0 @@
-// Package kafka package contains type definitions to describe the Apache Kafka transport in terms of Go-Kit abstractions.
-package kafka
