--- kafka/transport/doc.go
@@ -1,2 +1,0 @@
-// Package transport provides a Kafka transport abstraction.
-package transport
