+++ tracing/opentracing/propagation.go
@@ -4,13 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/alebabai/go-kafka"
+	"github.com/alebabai/go-kit-kafka/v2/transport"
+	"github.com/go-kit/log"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
-
-	"github.com/go-kit/log"
-
-	"github.com/alebabai/go-kit-kafka/kafka"
-	"github.com/alebabai/go-kit-kafka/kafka/transport"
 )
 
 // KafkaHeadersCarrier satisfies both [opentracing.TextMapWriter] and [opentracing.TextMapReader].
