+++ transport/doc.go
@@ -1,0 +1,2 @@
+// Package transport provides a Kafka transport abstractions.
+package transport
